+++ RLMod.Core/Infrastructure/Generator/MapGenerator.cs
@@ -262,84 +262,6 @@ public sealed class MapGenerator
         return values;
     }
 
-    public static class Validator
-    {
-        public class ValidatorResult
-        {
-            public bool IsConnected { get; set; }
-            public double ValueStdDev { get; set; }
-            public Dictionary<CountryType, int> CountryTypeDistribution { get; set; } = [];
-        }
-
-        public static ValidatorResult Validate(
-            IReadOnlyCollection<CountryMap> countries,
-            Dictionary<int, StateMap> stateMap
-        )
-        {
-            return new ValidatorResult
-            {
-                IsConnected = CheckConnectivity(countries, stateMap),
-                ValueStdDev = CalculateValueStdDev(countries),
-                CountryTypeDistribution = GetTypeDistribution(countries),
-            };
-        }
-
-        private static bool CheckConnectivity(
-            IEnumerable<CountryMap> countries,
-            Dictionary<int, StateMap> stateMap
-        )
-        {
-            foreach (var country in countries)
-            {
-                var visited = new HashSet<int>();
-                var queue = new Queue<int>();
-                queue.Enqueue(country.Id);
-
-                while (queue.Count > 0)
-                {
-                    int current = queue.Dequeue();
-                    if (!visited.Add(current))
-                    {
-                        continue;
-                    }
-
-                    foreach (
-                        int edge in stateMap[current]
-                            .Edges.AsValueEnumerable()
-                            .Where(edge => country.ContainsState(edge) || stateMap[edge].IsImpassable)
-                    )
-                    {
-                        if (!visited.Contains(edge))
-                        {
-                            queue.Enqueue(edge);
-                        }
-                    }
-                }
-
-                if (visited.Count != country.StateCount)
-                {
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        private static double CalculateValueStdDev(IEnumerable<CountryMap> countries)
-        {
-            double[] values = countries.AsValueEnumerable().Select(c => c.GetValue()).ToArray();
-            double mean = values.Average();
-            return Math.Sqrt(values.Average(v => Math.Pow(v - mean, 2)));
-        }
-
-        private static Dictionary<CountryType, int> GetTypeDistribution(IEnumerable<CountryMap> countries)
-        {
-            return countries
-                .AsValueEnumerable()
-                .GroupBy(c => c.Type)
-                .ToDictionary(g => g.Key, g => g.Count());
-        }
-    }
-
     public static class TestProgram
     {
         public static List<TmpState> GenerateTestStates(int count)
@@ -391,7 +313,7 @@ public sealed class MapGenerator
 
             var countries = generator.Divide();
 
-            var result = Validator.Validate(countries, CountryMap.StateMaps);
+            var result = MapValidator.Validate(countries, CountryMap.StateMaps);
 
             Console.WriteLine($"连通性验证: {result.IsConnected}");
             Console.WriteLine($"价值标准差: {result.ValueStdDev:F2}");
